feat: 너트워크 매니저 싱글톤 패턴으로 다시 수정
