+++ MovieReservation_4Team/Controller/MovieInfoViewController.swift
@@ -3,14 +3,14 @@ import UIKit
 class MovieInfoViewController: UIViewController {
     var movieInfoView = MovieInfoView()
     
-    let networkManager: NetworkManagerProtocol //같은 프로토콜 채택 중
-    init(networkManager: NetworkManagerProtocol) {
-        self.networkManager = networkManager
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
+//    let networkManager: NetworkManagerProtocol //같은 프로토콜 채택 중
+//    init(networkManager: NetworkManagerProtocol) {
+//        self.networkManager = networkManager
+//    }
+//    
+//    required init?(coder: NSCoder) {
+//        fatalError("init(coder:) has not been implemented")
+//    }
     
     override func loadView() {
         view = movieInfoView
