+++ MovieReservation_4Team/Model/Movie.swift
@@ -1,0 +1,20 @@
+//
+//  Movie.swift
+//  MovieReservation_4Team
+//
+//  Created by ahnzihyeon on 7/24/24.
+//
+
+import Foundation
+
+struct Movie: Decodable {
+    let id: Int
+    let title: String
+    let posterPath: String?
+    
+    enum CodingKeys: String, CodingKey {
+        case id
+        case title
+        case posterPath = "poster_path"
+    }
+}
