+++ MovieReservation_4Team/Model/NetworkManagerMock.swift
@@ -1,0 +1,29 @@
+//
+//  NetworkManagerMock.swift
+//  MovieReservation_4Team
+//
+//  Created by ahnzihyeon on 7/24/24.
+//
+
+import Foundation
+class NetworkManagerMock{
+//class NetworkManagerMock: NetworkManagerProtocol {
+//    func fetchPopularMovies(page: Int, completion: @escaping ([Movie]?) -> Void) {
+//        let movies: [Movie] = [ 
+//            .init(id: 1, title: "영화제목1", posterPath: <#T##String?#>),   //타입입을 지정해줘서 간단하게 .init으로 인스턴스 생성 가능
+//            .init(id: <#T##Int#>, title: <#T##String#>, posterPath: <#T##String?#>)  //이런 식으로 데이터 만들어주기
+//        ]
+//        
+//        completion(movies)
+//    }
+//    
+//    func fetchNowPlayingMovies(page: Int, completion: @escaping ([Movie]?) -> Void) {
+//        <#code#>
+//    }
+//    
+//    func fetchMovieDetail(movieId: Int, completion: @escaping (MovieDetail?) -> Void) {
+//        <#code#>
+//    }
+    
+    
+}
